pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "traceability-engine"
version = "0.1.0"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
description = "An AI-powered Git hook for automated method documentation and traceability."
readme = "README.md"
requires-python = ">=3.8"
keywords = ["git", "ai", "llama", "documentation", "ast"]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

# These are the libraries your tool needs to function
dependencies = [
    "openai",
    "GitPython",
    "python-dotenv"
]

[project.scripts]
# This creates the 'trace-gen' command in the user's terminal
# It maps 'trace-gen' to the main() function in traceability_engine/main.py
trace-gen = "traceability_engine.main:main"

[tool.setuptools.packages.find]
# Tells setuptools to look for the package logic in your folder
where = ["."]
include = ["traceability_engine*"]